moved mappers to separate class
