+++ core-services/src/test/java/org/zalando/nakadi/service/publishing/NakadiKpiPublisherTest.java
@@ -41,6 +41,7 @@ public class NakadiKpiPublisherTest {
     private final AvroSchema avroSchema = Mockito.mock(AvroSchema.class);
     private final UUIDGenerator uuidGenerator = Mockito.mock(UUIDGenerator.class);
     private final UsernameHasher usernameHasher = new UsernameHasher("123");
+    private final NakadiRecordMapper recordMapper = Mockito.mock(NakadiRecordMapper.class);
 
     @Captor
     private ArgumentCaptor<String> eventTypeCaptor;
@@ -56,7 +57,7 @@ public class NakadiKpiPublisherTest {
         final Supplier<JSONObject> dataSupplier = () -> null;
         new NakadiKpiPublisher(featureToggleService,
                 jsonProcessor, binaryProcessor, usernameHasher,
-                new EventMetadataTestStub(), uuidGenerator, avroSchema)
+                new EventMetadataTestStub(), uuidGenerator, avroSchema, recordMapper)
                 .publish("test_et_name", dataSupplier);
 
         verify(jsonProcessor).queueEvent("test_et_name", dataSupplier.get());
@@ -69,7 +70,7 @@ public class NakadiKpiPublisherTest {
         final Supplier<JSONObject> dataSupplier = () -> null;
         new NakadiKpiPublisher(featureToggleService,
                 jsonProcessor, binaryProcessor, usernameHasher,
-                new EventMetadataTestStub(), uuidGenerator, avroSchema)
+                new EventMetadataTestStub(), uuidGenerator, avroSchema, recordMapper)
                 .publish("test_et_name", dataSupplier);
 
         verify(jsonProcessor, Mockito.never()).queueEvent("test_et_name", dataSupplier.get());
@@ -85,7 +86,7 @@ public class NakadiKpiPublisherTest {
                 .setStatus("created");
 
         new NakadiKpiPublisher(featureToggleService, jsonProcessor, binaryProcessor, usernameHasher,
-                new EventMetadataTestStub(), uuidGenerator, avroSchema)
+                new EventMetadataTestStub(), uuidGenerator, avroSchema, recordMapper)
                 .publish(() -> subscriptionLogEvent);
 
         verify(jsonProcessor).queueEvent(eventTypeCaptor.capture(), jsonObjectCaptor.capture());
@@ -109,7 +110,7 @@ public class NakadiKpiPublisherTest {
         final Resource eventTypeRes = new DefaultResourceLoader().getResource("event-type-schema/");
         final var avroSchema = new AvroSchema(new AvroMapper(), new ObjectMapper(), eventTypeRes);
         new NakadiKpiPublisher(featureToggleService, jsonProcessor, binaryProcessor, usernameHasher,
-                new EventMetadataTestStub(), new UUIDGenerator(), avroSchema)
+                new EventMetadataTestStub(), new UUIDGenerator(), avroSchema, recordMapper)
                 .publish(() -> subscriptionLogEvent);
 
         verifyNoInteractions(jsonProcessor);
@@ -137,7 +138,7 @@ public class NakadiKpiPublisherTest {
         when(featureToggleService.isFeatureEnabled(Feature.KPI_COLLECTION)).thenReturn(false);
         final Supplier<KPIEvent> mockEventSupplier = Mockito.mock(Supplier.class);
         new NakadiKpiPublisher(featureToggleService, jsonProcessor, binaryProcessor, usernameHasher,
-                new EventMetadataTestStub(), uuidGenerator, avroSchema)
+                new EventMetadataTestStub(), uuidGenerator, avroSchema, recordMapper)
                 .publish(mockEventSupplier);
         verifyNoInteractions(mockEventSupplier, jsonProcessor, binaryProcessor, uuidGenerator, avroSchema);
     }
@@ -146,7 +147,7 @@ public class NakadiKpiPublisherTest {
     public void testHash() throws Exception {
         final NakadiKpiPublisher publisher = new NakadiKpiPublisher(featureToggleService,
                 jsonProcessor, binaryProcessor, usernameHasher,
-                new EventMetadataTestStub(), uuidGenerator, avroSchema);
+                new EventMetadataTestStub(), uuidGenerator, avroSchema, recordMapper);
         assertThat(publisher.hash("application"),
                 equalTo("befee725ab2ed3b17020112089a693ad8d8cfbf62b2442dcb5b89d66ce72391e"));
     }
@@ -160,7 +161,7 @@ public class NakadiKpiPublisherTest {
         final var avroSchema = new AvroSchema(new AvroMapper(), new ObjectMapper(), eventTypeRes);
         final NakadiKpiPublisher publisher = new NakadiKpiPublisher(featureToggleService,
                 jsonProcessor, binaryProcessor, usernameHasher,
-                new EventMetadataTestStub(), new UUIDGenerator(), avroSchema);
+                new EventMetadataTestStub(), new UUIDGenerator(), avroSchema, recordMapper);
         publisher.publishAccessLogEvent("POST",
                 "/test", "", "", "", "",
                 "", 200, 1L, 1L, 1L);
